requirements.txt
#Python 3.8+

requests>=2.31.0

beautifulsoup4>=4.12.0

openpyxl>=3.1.0
xlrd>=2.0.1

Pillow>=10.0.0

pytesseract>=0.3.10

pandas>=2.0.0

pyinstaller>=6.0.0